Update HomeScreen.js
Added some more conditionals to 'products && products.length && ...' to resolve 'TypeError: products.map is not a function' error which can occur if the request to Mongo Atlas for products takes longer than the speed of the page render. Every time I clicked on a product to view it's details, then clicked 'Go Back' or to the homepage I would get 'TypeError: products.map is not a function' consistently until I added the above changes.