Update Rating.js
This code smell was changed for better readability. What is happening is if text is truthful (meaning it is not null, undefined, 0, etc) then we render what the text prop actually is. Using the ternary operator we are able to read it much better as "If text is true and not null, then render text, otherwise, render nothing."