Fixed a bug where the save for later item was not being removed propeely
Fixed a bug where when the user logs out and logs back in with a DIFFERENT account, the same items in the "save for later" list were still there. Added to the userActions.js file to remove the later items list from the user's local storage when they log out.