Postfactor with comments explaining what I did + Set up stable coding environment As mentioned in the email to the Professor and TA, on 4/28/2023 I coded about 35% of feature 2 on Brian's apre linux laptop on my branch thinking I was signed into my git account. This is because my Windows computer is currently broken with WSL (path variables are destroyed and deleting/re-installing is not working). So Brian let me used his spare linux laptop and later that night we realized after he went home that it pushed under his account.
Fast foward to today, this morning I tried getting my commits in properly by copy/pasting the commits done in his account on the Windows computer, but this proved rather difficult as again, the project can't run on there. I made those commits around 3pm. That's when I decided to go to Best Buy to buy a MacBook with intents to return it after.

This leads to this specific commit. After purchasing the MacBook, re-buying Brad's Umedy course videos to re-learn how to set up from scratch, and getting a lot of help from my team mates this time as all of them have Macs and knew how to troubleshoot compared to when I was using my Windows computer, I was finally able to re set up the project just now. In this commit, I made comments explaining the 35% of feature 2 I coded on 4/28/2023 just so the GitHub can show my name and a proper git commit with working code.

Thank you in particular for your understanding and for Brian helping me the most when it came to setting up the mac book.